# Replace Stickyfill with native CSS sticky