--- a/sass/_sidebars.scss
+++ b/sass/_sidebars.scss
@@ -17,6 +17,7 @@
   }
 
   .sidebar-sticky {
+    position: -webkit-sticky;
     position: sticky;
     top: 1rem;
   }