--- a/templates/page_footer.html
+++ b/templates/page_footer.html
@@ -15,7 +15,6 @@
 <script type="text/javascript" src="https://use.fontawesome.com/a1f20be65b.js"></script>
 <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery-mousewheel/3.1.13/jquery.mousewheel.min.js"></script>
 <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jScrollPane/2.0.23/script/jquery.jscrollpane.min.js"></script>
-<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/stickyfill/1.1.4/stickyfill.min.js"></script>
 <script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/tooltipster/3.3.0/js/jquery.tooltipster.min.js"></script>
 {%- if md.global_config['enable_adsense'] %}
     <script type="text/javascript" src="{{ static_root }}js/merged-google.js?{{ now.timestamp() }}"></script>