Speeds up page load for virtually all users at the cost of a worse (but still usable) layout for the remaining ≤0.2% of IE users (according to Google Analytics browser stats provided in the Forge Discord server)