@LexManos 2022-05-21
I fully trust you on this stuff, but is there a chance of getting some profiling/stats to justify this stuff?
Obviously removing a external link will speed it up. But numbers are good.