@PaintNinja 2022-05-21
I'm testing on a fast desktop so there isn't a significant difference, but here's what I measured anyway in the Edge dev tools performance tab with cache disabled and a few runs on each:

| Metric    | Before | After | Difference |
| --------- | ------ | ----- | ---------- |
| Loading   | 11ms   | 10ms  | -9%        |
| Scripting | 66ms   | 60ms  | -9%        |
| Rendering | 62ms   | 57ms  | -8%        |
| Painting  | 35ms   | 33ms  | -6%        |
| System    | 90ms   | 84ms  | -7%        |