@PaintNinja 2022-05-21
We should probably get rid of [mousewheel](https://github.com/jquery/jquery-mousewheel/) too, that seems to be unused on the site (only usages I can find are loading the script, but not actually using it).
https://github.com/MinecraftForge/Web/blob/21733bec329c186a72c896458ba3083e312b6df9/templates/page_footer.html#L16